AbapGit Push : EMF Modeling
EMF Model for the ADT Abapgit Staging view plugin.

EMF Model is designed under the namespace "abapgitstaging" with mainly
four sections. Sections "UnstagedObjects", "IgnoredObjects" and
"StagedObjects" in the model are for holding the unstaged, ignored and
staged objects in the repo respectively. Each of these sections consist
of a list of abapgit objects and each abapgit object consist of a list
of files. The section "AbapGitCommitMessage" is for holding the commit
message of a commit which includes the message, author and the committer
details.