AbapGit Push (#154)
* AbapGit Push : EMF Modeling

EMF Model for the ADT Abapgit Staging view plugin.

EMF Model is designed under the namespace "abapgitstaging" with mainly
four sections. Sections "UnstagedObjects", "IgnoredObjects" and
"StagedObjects" in the model are for holding the unstaged, ignored and
staged objects in the repo respectively. Each of these sections consist
of a list of abapgit objects and each abapgit object consist of a list
of files. The section "AbapGitCommitMessage" is for holding the commit
message of a commit which includes the message, author and the committer
details.

* AbapGit Push : Generated Model Code

Generated model code from the EMF model.

* AbapGit Push : Content handler for EMF model

Content handler for the EMF model for handling back-end REST calls using
the model

* Extend the plug-in activator class from AbstractUIPlugin

Extend the plug-in activator class "AbapGitUIPlugin" with the
"AbstractUIPlugin" base class for more control over the plugin instance.

* AbapGit Push : Basic UI layout

Basic UI layout for the AbapGit Staging plugin.

AbapGit Staging view layout is inspired from the eclipse "Egit view" and
is mainly divided into three sections,

a) "Unstaged Changes" section : Lists all the repository objects which
are currently unstaged

b) Staged Changes section : List of all the repository objects which are
to be committed

c) Commit Message section : Consists of commit message, author and
committer details for the commit

* AbapGit Push : Add "stage object" and "unstage object" toolbar actions

"Stage object" toolbar action is provided for the unstaged changes
section for staging the selected objects in unstaged section tree viewer

"Unstage object" toolbar action is provided for the staged changes
section for unstaging the selected objects in staged section tree viewer

* AbapGit Push: Enable drag and drop of abap objects

Drag and drop support of objects between unstaged changes and staged
changes section.

* AbapGit Push: Enable drag and drop of abap objects

Drag and drop support of objects between unstaged changes and staged
changes section.

* AbabGit Push: Staging logic

Logic for refreshing the staging view and loading the contents from a
given abapgit repository.

* AbapGit Push: "Commit Message" section input validation logic

* Revert "AbapGit Push: "Commit Message" section input validation logic"

This reverts commit 5b4baefc8ec5bc891ffbc4c1d5dd04988df5fff2.

* AbapGit Push: "Commit Message" section UI input validation logic

* AbapGit Push: Commit Logic

- Commit Logic
- Dialog for github credentials for commit
- IExternalRepositoryInfoRequest is made public

* AbapGit Push: Stage and Commit atom link handling

Enhanced the repository serializer to support staging and commit atom
links

* Context menu action for opening AbapGit Staging view from repo view

* AbapGit Push: Util Package

- moved drag listeners to util package

* AbapGit Push: Github credentials handling

- Github account credentials handling for stage call, in case the repo
is private

- Github account credentials handling for commit call

Credentials are sent as part of the http request header.

* AbapGit Push: Adjust client to match with back-end + minor bug fixes

- Adjust the content type to match with the backend
- Exception handling for new stage and commit back-end calls. Outdated
client exception is not handled yet as it requires a dependency to the
ADT plugin "com.sap.adt.compatibility"
- Fixed an issue in handling username and password in Credentials dialog

* AbapGit Push: Icons for abap objects and files

- added a new attribute "wbkey" to abap object model which hold the
workbench object type value which will be used to display icons for the
object types
- icons for abap objects in unstaged and staged tree viewer
- icons for files in unstaged and staged tree viewer

* Rearrange the context menu actions in AbapGit repo view

* AbapGit Push: Refactoring + Bug fixes

- changes the type of localState and remoteState attributes to string
from char to avoid ascii conversion issues with EMF char type

- caching credentials for the selected repo : to avoid multiple
credentials dialog for each action

- UI text changes

- code refactoring : name/constants changes to adapt to bakend , refresh
logic refactoring , improved exception handling

- bug fix : changing the file status constant for DELETE status from 'R'
to 'D'

- response dialog for push

* AbapGit Push: Enable abap object navigation on double click

- navigation to the object adt editor on double click from unstaged or
staged section

* AbapGit Push: Context menu actions

- Open abap object
- Copy name
- Stage
- Unstage

Change-Id: I5c040fcc6b3b7212b06c37a429c09bcdb4e9848e

* AbapGit Push: Add tree comparator

* AbapGitPush: Tree filter text box and minor fixes

- new toolbar text filter box for tree filtering
- setting proper sort direction for tree comparator

* AbapGit Push: UI enhancements

- Retain the expanded state behavior of tree nodes on refresh, stage and
unstage action

- Open package toolbar action

- new icons

- collapse all button for unstaged and staged section

* AbapGit Push: Single file stage/unstage (drag/drop) mode impl

- enable single file drag and drop / stage and unstage support.
currently it is disabled and only enabled for non-code and meta files
node. Will enable it once the performance is tested during tree refresh
for a bigger repository.

* AbapGit Push: Adapt code for PDE tests + minor fixes

- Adapt the logic for writing PDE tests
- name changes + minor refactoring

* AbapGit Push: Handle Outdated Client exception

* AbapGit Push: Enhancements + fixes

- Project listener for project close and delete events for resetting the
view for such events

- disable project logon for test purposes

- disable commit button if no staged changes

- handle HTTP 409 exceptions which will be raised by backend for
parallel jobs

- adapt to push background job ( disable refresh after push, change
popup messages )

* AbapGit Push: Bug Fixes(spotbug) + refactoring

1. Fixes warnings reported from spotbugs
2. new icons
3. refactoring
4. messages changes after review
5. added mnemonics
6. always read credentials during push, because since push is now a
background job it is not possible to validate and get the exception in
the client if the credentials which user entered is invalid. So it is
not possible to cache the credentials for commit.

* AbapGit Push : PDE Tests

* AbapGit Push: Changing icon + adding the svg file for the icon

* AbapGit Push: Resolve maven build issues and spotbug issues

1. Reduce the dependency on com.sap.adt.compatibility plugin to 2.106.1
version

2. Add the dependency on easymock in the parent POM

3. Solve spotbugs issues

4. Remove spotbugs checks on generated emf files