[abapGit] Encode file handling URL's before the backend call
Encode the URL for retrieving file contents before the backend call.
This is to prevent errors if the file name contains any special
characters.

From 2011 release onwards, backend sends the encoded URL to the client,
so the client does not need to handle the encoding of file names. But
this change is required for compatibility reasons to support 2008
backend.