revert the changes for improved handling  transport page in link wizard
In case a pull action is not performed while linking a repository,
the transport page is skipped in the link a repository wizard.
Currently, a back-end check happens for transport request for the
package provided while linking the repository, which breaks the client.
This check is not required.

Thus until the check is removed from the back end, the changes related
to skipping transport page if a link is performed without pull, have to
be reverted.