Disable selective pull feature temporarily
Backend call to fetch locally modified objects in a repository during
PULL is running into an error which results in PULL action being
aborted.

Happens while linking and pulling existing abapGit repositories which
also contains a sub-package.

Reason for the failure:
Selective pull backend call would try to create sub-packages in the
system which are part of the repository and this would require a
transport request which is currently not being passed.

A proper fix would be to handle the transport request with selective
pull backend call which would be done as a separate fix.